+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - TBD
+
+### Changed
+
+### Added
+
+- Support for block iterator for the `LDMatrix`.
+- `rank_one_update` for the `LDLinearOperator` class.
+- Unified method to map variants to genomic blocks `map_variants_to_genomic_blocks`.
 
 ## [0.1.4] - 2024-10-01